Changed Int to double in division method

+++ Kalkylator/Program.cs
@@ -13,11 +13,10 @@
             return a * b;
         }
 
-
-        static double Division(int a, int b)
-        {
-            return a / b;
-        }
+        static double Division(double a, double b)
+            {
+              return a / b;
+            }
 
 
         static int Addition(int a, int b)
